Added dto layer for book records. Fixed ids with BIGINT fields in DB

+++ sql/01-schema/01_book_record_table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE book_record
 (
-    book_id SERIAL NOT NULL,
+    book_id BIGINT NOT NULL,
     title   VARCHAR(255) NOT NULL,
     author  VARCHAR(255) NOT NULL,
     price   NUMERIC(10,2) NOT NULL,
